backend/requirements.txt
Flask
flask-cors
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
python-dotenv
openai
sqlalchemy
bcrypt
pydantic
requests
werkzeug
gunicorn
flask-migrate